icon-previous: added button usage to docs
