+++ icon/doc/components/ButtonPreviewContainer.styles.tsx
@@ -1,0 +1,6 @@
+import { styled } from '@stitches/react';
+
+export const ButtonPreviewContainer = styled('div', {
+  display: 'flex',
+  flexDirection: 'row',
+});
