+++ icon/doc/icon.mdx
@@ -2,8 +2,10 @@ import { mdx } from '@mdx-js/react';
 import { MdxButtonLayout } from '~/layout';
 import { Props, Description } from '@divriots/dockit-react/props';
 import { Playground } from '@divriots/dockit-react/playground';
+import CheckIcon from '../icons/CheckIcon/CheckIcon';
+import Button from '../../Button/src/Button'
 import Icon from '../src/Icon';
-import IconPreview from './components/IconPreview';
+import { ButtonPreviewContainer } from './components/ButtonPreviewContainer.styles';
 export default MdxButtonLayout;
 
 # Custom Icon
@@ -13,27 +15,37 @@ export default MdxButtonLayout;
 ### No Overridden Styling
 Icons have default styling for SVG styling like fill, stroke, etc. These are defaulted to brand values when no extra props are passed in.
 <Playground
-  scope={{ Icon }}
+  scope={{ CheckIcon }}
   code={`
-<Icon icon="home" size="s" />
-<Icon icon="home" size="m" />
-<Icon icon="home" size="l" />
-<Icon icon="home" size="xl" />
+<CheckIcon size="s" />
+<CheckIcon size="m" />
+<CheckIcon size="l" />
+<CheckIcon size="xl" />
   `}
 />
 
 ### Overridden Styling
-Styling can be overridden if required as well.
+Fill and color can be passed in as props.
 <Playground
-  scope={{ Icon }}
+  scope={{ CheckIcon }}
   code={`
-    <Icon icon="home"  fill="#F0BD6B" stroke="#D17F13" />
+      <CheckIcon size="m" color="red" fill="purple" />
   `}
+/>
+
+### Usage With Other Components
+The Icon components can be used inside other components as well, e.g. Button Component
+<Playground
+  scope={{ CheckIcon, Button, ButtonPreviewContainer }}
+  code={`
+  <ButtonPreviewContainer>
+         <Button size="m" icon={<CheckIcon />}>Submit</Button>
+        <Button size="m" icon={<CheckIcon />} iconPlacement="left">Submit</Button>
+        <Button size="m" icon={<CheckIcon />} />
+  </ButtonPreviewContainer>
+`}
 />
 
 ## Props
 We expose a subset of SVGProps to our component, along with some custom props
-<Props of={Icon} />
-
-## Icons
-<IconPreview />
+<Props of={CheckIcon} />
