+++ icon/stories/Icon.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import Icon from '../src/Icon';
+import CheckIcon from '../icons/CheckIcon/CheckIcon';
 import { layout } from '../../utils/src/story-decorator';
 export default layout;
 
-export const IconSmall = () => <Icon icon="home" size="s" />;
-export const IconMedium = () => <Icon icon="home" size="m" />;
-export const IconLarge = () => <Icon icon="home" size="l" />;
-export const IconExtraLarge = () => <Icon icon="home" size="xl" />;
+export const IconSmall = () => <CheckIcon size="s" color="red" />;
+export const IconMedium = () => <CheckIcon />;
+export const IconLarge = () => <CheckIcon size="l" color="blue" />;
+export const IconExtraLarge = () => <CheckIcon size="xl" fill="green" color="pink" />;
