+++ icon/types/index.ts
@@ -1,0 +1,8 @@
+import { SVGProps } from 'react';
+
+export type Sizes = 's' | 'm' | 'l' | 'xl';
+
+export type PickedSVGProps = Pick<
+  SVGProps<SVGPathElement>,
+  'fill' | 'stroke' | 'strokeLinecap' | 'strokeLinejoin' | 'strokeWidth' | 'transform' | 'd'
+>;
